Move <ScrollUpButton /> to `interactions` module
